webapps/dashboard/requirements.txt
asgiref==3.11.0
Django==6.0
sqlparse==0.5.5
uWSGI==2.0.31
djangorestframework==3.16.1

# requests and its dependencies
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
requests==2.32.5
urllib3==2.6.3
